Add 12 CPU modules + tests; fix 10 pre-existing bugs
MGARD/SPERR fused paths kept as *FusedDecomposition and still drive
ALGO_MGARD / ALGO_SPERR; the composable compositions take the default names.

+++ include/SZ3/encoder/ArithmeticEncoder.hpp
@@ -533,7 +533,7 @@ class ArithmeticEncoder : public concepts::EncoderInterface<T> {
         size_t total_frequency = ariCoder.total_frequency;
         const uchar *sp = bytes + 5;
         unsigned int offset = 4;
-        size_t value = (bytesToInt64_bigEndian(bytes) >> 20);  // alignment with the MAX_CODE
+        size_t value = (static_cast<uint64_t>(bytesToInt64_bigEndian(bytes)) >> 20);  // alignment with the MAX_CODE
         size_t s_counter = sizeof(int);
 
         for (i = 0; i < targetLength; i++) {
